Pass the logic of get the login info into a hook
